Remove Gemfile.lock
GitHub controls versions of Gems used and upgrades them
on a schedule out of my control. Recording the Gemfile.lock
was a way to ensure local development was using the last
known good version. However, the Gemfile gets the latest versions
of Ruby and github-pages from GitHub, so this is unnecessary.

Removing Gemfile.lock will prevent vulnerability warnings that
are false negatives.